Divide classes into packages
The packages are named bond.main/.task/.command
