+++ src/main/java/bond/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package command;
+
+import bond.BondException;
+import bond.Storage;
+import bond.Ui;
+import task.Task;
+import task.TaskList;
+
 public class DeleteCommand extends Command {
 
     private int index;
