+++ src/main/java/bond/main/Storage.java
@@ -1,3 +1,5 @@
+package bond;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -5,6 +7,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.ListIterator;
 import java.util.Scanner;
+import task.*;
 
 public class Storage {
