+++ src/main/java/bond/task/EventTask.java
@@ -1,5 +1,9 @@
+package bond.task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+
+import bond.main.BondException;
 
 public class EventTask extends Task {
