+++ src/main/java/bond/task/Task.java
@@ -1,3 +1,5 @@
+package task;
+
 public abstract class Task {
 
     private Boolean isCompleted;
